server/requirements.txt
Django==3.2.7
django-cors-headers==3.8.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
mysqlclient==2.0.3
django-allauth==0.45.0
django-rest-swagger==2.2.0
dj-rest-auth==2.1.11